kLdrModLX.c: warnings
git-svn-id: http://svn.netlabs.org/repos/kstuff/trunk@96 815646eb-5b94-dc11-ae90-00e081727c95

+++ kLdr/kLdrModLX.c
@@ -1522,7 +1522,7 @@ static int kldrModLXDoIterDataUnpacking(KU8 *pbDst, const KU8 *pbSrc, int cbSrc)
     int                     cbDst = OBJPAGELEN;
 
     /* Validate size of data. */
-    if (cbSrc >= OBJPAGELEN - 2)
+    if (cbSrc >= (int)OBJPAGELEN - 2)
         return KLDR_ERR_LX_BAD_ITERDATA;
 
     /*
@@ -2474,11 +2474,11 @@ static int kldrModLXRelocateBits(PKLDRMOD pMod, void *pvBits, KLDRADDR NewBaseAd
                     /* common / simple */
                     if (    (u.prlc->nr_stype & NRSRCMASK) == NROFF32
                         &&  off >= 0
-                        &&  off <= OBJPAGELEN - 4)
+                        &&  off <= (int)OBJPAGELEN - 4)
                         *(KU32 *)&pbPage[off] = (KU32)uValue;
                     else if (    (u.prlc->nr_stype & NRSRCMASK) == NRSOFF32
                             &&  off >= 0
-                            &&  off <= OBJPAGELEN - 4)
+                            &&  off <= (int)OBJPAGELEN - 4)
                         *(KU32 *)&pbPage[off] = (KU32)(uValue - (PageAddress + off + 4));
                     else
                     {
@@ -2499,7 +2499,7 @@ static int kldrModLXRelocateBits(PKLDRMOD pMod, void *pvBits, KLDRADDR NewBaseAd
                         while (c-- > 0)
                         {
                             int off = *poffSrc++;
-                            if (off >= 0 && off <= OBJPAGELEN - 4)
+                            if (off >= 0 && off <= (int)OBJPAGELEN - 4)
                                 *(KU32 *)&pbPage[off] = (KU32)uValue;
                             else
                             {
@@ -2514,7 +2514,7 @@ static int kldrModLXRelocateBits(PKLDRMOD pMod, void *pvBits, KLDRADDR NewBaseAd
                         while (c-- > 0)
                         {
                             int off = *poffSrc++;
-                            if (off >= 0 && off <= OBJPAGELEN - 4)
+                            if (off >= 0 && off <= (int)OBJPAGELEN - 4)
                                 *(KU32 *)&pbPage[off] = (KU32)(uValue - (PageAddress + off + 4));
                             else
                             {
@@ -2647,7 +2647,7 @@ static int kldrModLXDoReloc(KU8 *pbPage, int off, KLDRADDR PageAddress, const st
     pbDst = pbPage + off;
     while (cb-- > 0)
     {
-        if (off > OBJPAGELEN)
+        if (off > (int)OBJPAGELEN)
             break;
         if (off >= 0)
             *pbDst = *pbSrc;
